Remove redirect of all dynamic pages to 404 and add dynamic page for displaying stream
