+++ apps/web/src/app/stream/sandbox/[sandboxId]/page.tsx
@@ -1,0 +1,51 @@
+import { Suspense } from 'react'
+import { createClient } from '@supabase/supabase-js'
+import MuxPlayer from '@mux/mux-player-react'
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!
+const supabaseServiceKey = process.env.SUPABASE_SERVICE_ROLE_KEY!
+
+const supabase = createClient(supabaseUrl, supabaseServiceKey)
+
+interface SandboxStream {
+  sandboxId: string
+  playbackId: string
+}
+
+async function fetchStream(sandboxId: string): Promise<SandboxStream | null> {
+  const { data, error } = await supabase
+    .from('sandbox_streams')
+    .select('playback_id')
+    .eq('sandbox_id', sandboxId)
+    .single()
+
+  if (error || !data) {
+    return null
+  }
+
+  return { sandboxId, playbackId: data.playback_id }
+}
+
+export default async function StreamPage({ params }: { params: { sandboxId: string } }) {
+  const stream = await fetchStream(params.sandboxId)
+
+  if (!stream) {
+    return <div>Stream not found</div>
+  }
+
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <h1>Stream Details</h1>
+      <div>Sandbox ID: {stream.sandboxId}</div>
+      <div>Playback ID: {stream.playbackId}</div>
+
+      <MuxPlayer
+        playbackId={stream.playbackId}
+        metadata={{
+          video_id: `sandbox-${stream.sandboxId}`,
+          video_title: 'Desktop Sandbox Stream',
+        }}
+      />
+    </Suspense>
+  )
+}
