+++ apps/web/src/types/mux-player-react.d.ts
@@ -1,0 +1,9 @@
+declare module '@mux/mux-player-react' {
+  export interface MuxPlayerProps {
+    [key: string]: any;
+    playbackId: string;
+    metadata?: Record<string, any>;
+  }
+
+  export default function MuxPlayer(props: MuxPlayerProps): JSX.Element
+}
